ENH: Add dynamic range keywords.
Assuming the dynamic range parameter would often be used like it is in the example, this would be a nice feature. On the other hand, the use in the example doesn't really have much of an effect on the output.

Also note that the use of dynamic range was buggy in the example (`img_noise` and `img_const` were switched in min/max calculation).