Implement lazy submodule importing
This allows scikit-image to be used as follows:

```
import skimage as ski
ski.filters.gaussian(...)
```

The lazy loading strategy could also be implemented inside submodules,
to speed up imports of that submodule.  E.g., we could do:

```
sla = lazy.require('scipy.linalg')
```

The linalg module is only loaded once used in the code, avoiding the
standard workaround of having to move the import inside of a function
call.